chore: added TSDocs inside of Modal component
