+++ packages/tradershub/src/components/Modal/ModalFooter.tsx
@@ -1,29 +1,26 @@
 import React, { ReactNode } from 'react';
-import { cva, VariantProps } from 'class-variance-authority';
-import { ExcludeAllNull, qtMerge } from '@deriv/quill-design';
+import { qtMerge } from '@deriv/quill-design';
+import { ModalFooterClass, TModalFooterClass } from './Modal.classnames';
 
-type TModalFooterClass = ExcludeAllNull<VariantProps<typeof ModalFooterClass>>;
-
-interface TModalFooter extends TModalFooterClass {
+/**
+ * Interface for the ModalFooter component props
+ * @interface TModalFooter
+ * @extends {TModalFooterClass}
+ * @property {ReactNode} children - Children nodes
+ * @property {string} [className] - Optional CSS class name
+ */
+export interface TModalFooter extends TModalFooterClass {
     children: ReactNode;
     className?: string;
 }
 
+/**
+ * ModalFooter component
+ * @param {TModalFooter} props - The properties that define the ModalFooter component.
+ * @returns {JSX.Element} The ModalFooter component.
+ */
 const ModalFooter = ({ align = 'right', children, className }: TModalFooter) => (
     <div className={qtMerge(ModalFooterClass({ align }), className)}>{children}</div>
 );
 
 export default ModalFooter;
-
-export const ModalFooterClass = cva(
-    'grid grid-cols-2 gap-400 p-800 border border-solid border-t-100 border-system-light-secondary-background bottom-0 lg:flex lg:items-center lg:px-1200 lg:py-800',
-    {
-        variants: {
-            align: {
-                center: 'lg:justify-center',
-                left: 'lg:justify-start',
-                right: 'lg:justify-end',
-            },
-        },
-    }
-);
